Bug fix in image normalization
We first multiplied and then divided. This caused data stored in ndarrays of limited size (in this case 16 bits) to be truncated, and data lost.

If instead we first divide and then multiply, the ndarray is coerced to a valid float array which stores all the data. Afterwards, we multiply and cast to uint8 making the data valid.